Converting 'var' into 'let' and 'const'.
Summary: While 'var' can be used anywhere once declared, this is not true for 'let' and 'const'.
They are only valid within their declared scope. This is useful so future developers won't
have to wonder where vars are being used, potentially in a different file with the except
of global variables like ActiveState. 'Const' was used wherever possible. This contains
the added benefit over 'let' and 'var' of being immutable (reference-wise). Now devs won't
have to wonder which variables change. Lastly, as ESLint was used to streamline the process
of changing 'var'/'let' to 'const', this commit includes some trivial edits that ESLint
alerted me to like an 'unnecessary backslash in regexp expression' or an 'unused parameter'.

Test: Did basic sanity testing on a localhost'd web server using previously saved hierarchy
data, a bug report, and using a live device.