Remove old time lapse file processing code.
This simplifies the maintenance costs so that we don't need to
consider legacy bug report time lapse data when making new changes.

There also were some bugs that required time to root cause, and it
was decided it was faster and simpler to just remove the code that
needed to be updated in the first place.